Eliminate clang-tidy warning and optimize headers

+++ cpp/abplas/base/lattice.cpp
@@ -10,7 +10,7 @@ namespace base {
 
 void frac2cart(const Eigen::Matrix3d &latticeVectors, const Eigen::Matrix3Xd &fracCoordinates,
                Eigen::Matrix3Xd &cartCoordinates) {
-    int num_col = fracCoordinates.cols();
+    long num_col = fracCoordinates.cols();
     cartCoordinates = Eigen::Matrix3Xd(3, num_col);
     for(int idx = 0; idx < num_col; idx++) {
         cartCoordinates.col(idx) = latticeVectors * fracCoordinates.col(idx);
@@ -19,7 +19,7 @@ void frac2cart(const Eigen::Matrix3d &latticeVectors, const Eigen::Matrix3Xd &fr
 
 void cart2frac(const Eigen::Matrix3d &latticeVectors, const Eigen::Matrix3Xd &cartCoordinates,
                Eigen::Matrix3Xd &fracCoordinates) {
-    int num_col = cartCoordinates.cols();
+    long num_col = cartCoordinates.cols();
     fracCoordinates = Eigen::Matrix3Xd(3, num_col);
     Eigen::Matrix3d conv_mat = latticeVectors.inverse();
     for(int idx = 0; idx < num_col; idx++) {
